Use conditional imports to not break non-web platforms
