+++ flutter_local_notifications/lib/src/web_mock.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_local_notifications_platform_interface/flutter_local_notifications_platform_interface.dart';
+
+/// A stub implementation of the web plugin, for non-web platforms.
+class WebFlutterLocalNotificationsPlugin extends FlutterLocalNotificationsPlatform {
+  Future<bool?> initialize() async { return null; }
+
+  Future<bool> requestNotificationsPermission() async { return false; }
+}
